+++ tests/server.js
@@ -1,6 +1,12 @@
 var ServerController = require('./../src/index.js').ServerController
 var suite = {}
 
+suite['should get top level'] = function (test) {
+  var ctrl = Controller.ServerController({foo: 'bar'})
+  test.deepEqual(ctrl.get(), {foo: 'bar'})
+  test.done()
+}
+
 suite['should get top level key'] = function (test) {
   var ctrl = ServerController({foo: 'bar'})
   test.equal(ctrl.get('foo'), 'bar')